Move event base WPs out of legacy
