+++ indico/modules/events/abstracts/views.py
@@ -19,9 +19,9 @@ from __future__ import unicode_literals
 from flask import render_template, session
 
 from indico.legacy.webinterface.pages.base import WPJinjaMixin
-from indico.legacy.webinterface.pages.conferences import WPConferenceDefaultDisplayBase
 from indico.modules.events.abstracts.util import get_visible_reviewed_for_tracks
 from indico.modules.events.management.views import WPEventManagement
+from indico.modules.events.views import WPConferenceDisplayLegacyBase
 from indico.util.mathjax import MathjaxMixin
 
 
@@ -43,24 +43,24 @@ class WPManageAbstracts(MathjaxMixin, WPJinjaMixin, WPEventManagement):
         return WPEventManagement._getHeadContent(self) + MathjaxMixin._getHeadContent(self)
 
 
-class WPDisplayAbstractsBase(MathjaxMixin, WPJinjaMixin, WPConferenceDefaultDisplayBase):
+class WPDisplayAbstractsBase(MathjaxMixin, WPJinjaMixin, WPConferenceDisplayLegacyBase):
     template_prefix = 'events/abstracts/'
 
     def getJSFiles(self):
-        return (WPConferenceDefaultDisplayBase.getJSFiles(self) +
+        return (WPConferenceDisplayLegacyBase.getJSFiles(self) +
                 self._asset_env['markdown_js'].urls() +
                 self._asset_env['selectize_js'].urls() +
                 self._asset_env['modules_reviews_js'].urls() +
                 self._asset_env['modules_abstracts_js'].urls())
 
     def getCSSFiles(self):
-        return WPConferenceDefaultDisplayBase.getCSSFiles(self) + self._asset_env['selectize_css'].urls()
+        return WPConferenceDisplayLegacyBase.getCSSFiles(self) + self._asset_env['selectize_css'].urls()
 
     def _getBody(self, params):
         return WPJinjaMixin._getPageContent(self, params).encode('utf-8')
 
     def _getHeadContent(self):
-        return WPConferenceDefaultDisplayBase._getHeadContent(self) + MathjaxMixin._getHeadContent(self)
+        return WPConferenceDisplayLegacyBase._getHeadContent(self) + MathjaxMixin._getHeadContent(self)
 
 
 class WPDisplayAbstracts(WPDisplayAbstractsBase):
