+++ indico/modules/events/registration/views.py
@@ -17,10 +17,9 @@
 from __future__ import unicode_literals
 
 from indico.legacy.webinterface.pages.base import WPJinjaMixin
-from indico.legacy.webinterface.pages.conferences import WPConferenceDefaultDisplayBase
 from indico.modules.events.management.views import WPEventManagement
 from indico.modules.events.models.events import EventType
-from indico.modules.events.views import WPSimpleEventDisplayBase
+from indico.modules.events.views import WPConferenceDisplayLegacyBase, WPSimpleEventDisplayBase
 
 
 class WPManageRegistration(WPEventManagement):
@@ -68,9 +67,9 @@ class DisplayRegistrationFormMixin(WPJinjaMixin):
         return self.base_class.getJSFiles(self) + self._asset_env['modules_registration_js'].urls()
 
 
-class WPDisplayRegistrationFormConference(DisplayRegistrationFormMixin, WPConferenceDefaultDisplayBase):
+class WPDisplayRegistrationFormConference(DisplayRegistrationFormMixin, WPConferenceDisplayLegacyBase):
     template_prefix = 'events/registration/'
-    base_class = WPConferenceDefaultDisplayBase
+    base_class = WPConferenceDisplayLegacyBase
     menu_entry_name = 'registration'
