+++ indico/modules/events/reminders/views.py
@@ -17,10 +17,10 @@
 from __future__ import unicode_literals
 
 from indico.legacy.webinterface.pages.base import WPJinjaMixin
-from indico.legacy.webinterface.pages.conferences import WPConferenceModifBase
+from indico.modules.events.management.views import WPEventManagementLegacy
 
 
-class WPReminders(WPConferenceModifBase, WPJinjaMixin):
+class WPReminders(WPEventManagementLegacy, WPJinjaMixin):
     template_prefix = 'events/reminders/'
     sidemenu_option = 'reminders'
