+++ indico/modules/events/timetable/views/__init__.py
@@ -25,12 +25,12 @@ from sqlalchemy.orm import joinedload
 
 from indico.core import signals
 from indico.legacy.webinterface.pages.base import WPJinjaMixin
-from indico.legacy.webinterface.pages.conferences import WPConferenceDefaultDisplayBase
 from indico.modules.events.layout import theme_settings
 from indico.modules.events.management.views import WPEventManagement
 from indico.modules.events.timetable.models.entries import TimetableEntryType
 from indico.modules.events.timetable.views.weeks import inject_week_timetable
 from indico.modules.events.util import get_theme
+from indico.modules.events.views import WPConferenceDisplayLegacyBase
 from indico.util.signals import values_from_signal
 from indico.web.flask.templating import register_template_hook, template_hook
 
@@ -51,7 +51,7 @@ class WPManageTimetable(WPEventManagement):
                 self._asset_env['modules_contributions_js'].urls())
 
 
-class WPDisplayTimetable(WPJinjaMixin, WPConferenceDefaultDisplayBase):
+class WPDisplayTimetable(WPJinjaMixin, WPConferenceDisplayLegacyBase):
     template_prefix = 'events/timetable/'
     menu_entry_name = 'timetable'
 
@@ -59,7 +59,7 @@ class WPDisplayTimetable(WPJinjaMixin, WPConferenceDefaultDisplayBase):
         return WPJinjaMixin._getPageContent(self, params)
 
     def getJSFiles(self):
-        return WPConferenceDefaultDisplayBase.getJSFiles(self) + self._asset_env['modules_timetable_js'].urls()
+        return WPConferenceDisplayLegacyBase.getJSFiles(self) + self._asset_env['modules_timetable_js'].urls()
 
 
 @template_hook('meeting-body')
