+++ indico/modules/events/views.py
@@ -14,26 +14,119 @@
 # You should have received a copy of the GNU General Public License
 # along with Indico; if not, see <http://www.gnu.org/licenses/>.
 
-from __future__ import unicode_literals
+from __future__ import print_function, unicode_literals
 
+import os
 import posixpath
 
-from flask import render_template, request
+from flask import render_template, render_template_string, request
 from sqlalchemy.orm import load_only
 
+from indico.core.config import config
 from indico.legacy.webinterface.pages.base import WPDecorated, WPJinjaMixin
-from indico.legacy.webinterface.pages.conferences import (WPConferenceDefaultDisplayBase, WPEventBase,
-                                                          render_event_footer, render_event_header)
 from indico.modules.admin.views import WPAdmin
+from indico.modules.core.settings import core_settings, social_settings
 from indico.modules.events import Event
 from indico.modules.events.layout import layout_settings, theme_settings
-from indico.modules.events.layout.util import get_css_url
+from indico.modules.events.layout.util import (build_menu_entry_name, get_css_url, get_menu_entry_by_name,
+                                               menu_entries_for_event)
+from indico.modules.events.models.events import EventType
+from indico.util.date_time import format_date
 from indico.util.event import unify_event_args
 from indico.util.mathjax import MathjaxMixin
+from indico.util.string import strip_tags, to_unicode, truncate
+from indico.web.flask.util import url_for
+
+
+def _get_print_url(event, theme=None, theme_override=False):
+    view = theme if theme_override else None
+
+    if event.type_ == EventType.conference:
+        return url_for('timetable.timetable', event, print='1', view=view)
+    elif event.type_ == EventType.meeting:
+        show_date = request.args.get('showDate')
+        show_session = request.args.get('showSession')
+        detail_level = request.args.get('detailLevel')
+        if show_date == 'all':
+            show_date = None
+        if show_session == 'all':
+            show_session = None
+        if detail_level in ('all', 'contrinbution'):
+            detail_level = None
+        return url_for('events.display', event, showDate=show_date, showSession=show_session, detailLevel=detail_level,
+                       print='1', view=view)
+    elif event.type_ == EventType.lecture:
+        return url_for('events.display', event, print='1', view=view)
+
+
+def render_event_header(event, conference_layout=False, theme=None, theme_override=False):
+    print_url = _get_print_url(event, theme, theme_override) if not conference_layout else None
+    show_nav_bar = event.type_ != EventType.conference or layout_settings.get(event, 'show_nav_bar')
+    themes = {tid: {'name': data['title'], 'user_visible': data.get('user_visible')}
+              for tid, data in theme_settings.get_themes_for(event.type_.name).viewitems()}
+    return render_template('events/header.html',
+                           event=event, print_url=print_url, show_nav_bar=show_nav_bar, themes=themes, theme=theme)
+
+
+def render_event_footer(event, dark=False):
+    location = event.venue_name
+    if event.room_name:
+        location = '{} ({})'.format(event.room_name, location)
+    description = '{}\n\n{}'.format(truncate(event.description, 1000), event.short_external_url).strip()
+    google_calendar_params = {
+        'action': 'TEMPLATE',
+        'text': event.title,
+        'dates': '{}/{}'.format(event.start_dt.strftime('%Y%m%dT%H%M%SZ'),
+                                event.end_dt.strftime('%Y%m%dT%H%M%SZ')),
+        'details': description,
+        'location': location,
+        'trp': False,
+        'sprop': [event.external_url, 'name:indico']
+    }
+
+    social_settings_data = social_settings.get_all()
+    show_social = social_settings_data['enabled'] and layout_settings.get(event, 'show_social_badges')
+    return render_template('events/footer.html',
+                           event=event,
+                           dark=dark,
+                           social_settings=social_settings_data,
+                           show_social=show_social,
+                           google_calendar_params=google_calendar_params)
 
 
 class WPReferenceTypes(WPAdmin):
     template_prefix = 'events/'
+
+
+class WPEventBase(WPDecorated):
+    @unify_event_args
+    def __init__(self, rh, event_, **kwargs):
+        assert event_ == kwargs.setdefault('event', event_)
+        self.event = event_
+        WPDecorated.__init__(self, rh, **kwargs)
+        self._navigationTarget = self._conf = event_.as_legacy
+        self._tz = event_.display_tzinfo.zone
+        start_dt_local = event_.start_dt_display.astimezone(event_.display_tzinfo)
+        end_dt_local = event_.end_dt_display.astimezone(event_.display_tzinfo)
+        dates = ' ({})'.format(to_unicode(format_date(start_dt_local, format='long')))
+        if start_dt_local.date() != end_dt_local.date():
+            if start_dt_local.year == end_dt_local.year and start_dt_local.month == end_dt_local.month:
+                dates = ' ({}-{})'.format(start_dt_local.day, to_unicode(format_date(end_dt_local, format='long')))
+            else:
+                dates = ' ({} - {})'.format(to_unicode(format_date(start_dt_local, format='long')),
+                                            to_unicode(format_date(end_dt_local, format='long')))
+        self._setTitle('{} {}'.format(strip_tags(self.event.title), dates))
+
+    def _getHeader(self):
+        raise NotImplementedError  # must be overridden by meeting/lecture and conference WPs
+
+    def getJSFiles(self):
+        return WPDecorated.getJSFiles(self) + self._asset_env['modules_event_display_js'].urls()
+
+    def _getHeadContent(self):
+        site_name = core_settings.get('site_title')
+        meta = render_template('events/meta.html', event=self.event, site_name=site_name)
+        return WPDecorated._getHeadContent(self) + meta
 
 
 class WPSimpleEventDisplayBase(MathjaxMixin, WPEventBase):
@@ -120,14 +213,78 @@ class WPSimpleEventDisplay(WPSimpleEventDisplayBase):
         return rv.encode('utf-8')
 
 
-class WPAccessKey(WPJinjaMixin, WPDecorated):
-    template_prefix = 'events/'
+class WPConferenceDisplayLegacyBase(MathjaxMixin, WPEventBase):
+    menu_entry_plugin = None
+    menu_entry_name = None
 
-    def _getBody(self, params):
-        return self._getPageContent(params)
+    @unify_event_args
+    def __init__(self, rh, event_, **kwargs):
+        assert event_ == kwargs.setdefault('event', event_)
+        self.event = event_
+        kwargs['conf_layout_params'] = self._get_layout_params()
+        WPEventBase.__init__(self, rh, event_, **kwargs)
+
+    def _get_layout_params(self):
+        bg_color = layout_settings.get(self.event, 'header_background_color').replace('#', '').lower()
+        text_color = layout_settings.get(self.event, 'header_text_color').replace('#', '').lower()
+        announcement = ''
+        if layout_settings.get(self.event, 'show_announcement'):
+            announcement = layout_settings.get(self.event, 'announcement')
+        return {
+            'menu': menu_entries_for_event(self.event),
+            'active_menu_item': self.sidemenu_option,
+            'bg_color_css': 'background: #{0}; border-color: #{0};'.format(bg_color) if bg_color else '',
+            'text_color_css': 'color: #{};'.format(text_color) if text_color else '',
+            'announcement': announcement
+        }
+
+    def get_extra_css_files(self):
+        theme_url = self._kwargs.get('css_url_override', get_css_url(self._conf.as_event))
+        return [theme_url] if theme_url else []
+
+    def _getHeader(self):
+        return render_event_header(self.event, conference_layout=True).encode('utf-8')
+
+    @property
+    def sidemenu_option(self):
+        if not self.menu_entry_name:
+            return None
+        name = build_menu_entry_name(self.menu_entry_name, self.menu_entry_plugin)
+        entry = get_menu_entry_by_name(name, self.event)
+        if entry:
+            return entry.id
+
+    def _apply_conference_layout(self, body):
+        tpl = "{% extends 'events/display/conference/base.html' %}{% block content %}{{ _body | safe }}{% endblock %}"
+        return render_template_string(tpl, _body=body, **self._kwargs)
+
+    def _getHeadContent(self):
+        path = config.BASE_URL
+        try:
+            timestamp = os.stat(__file__).st_mtime
+        except OSError:
+            timestamp = 0
+        css = '<link rel="stylesheet" type="text/css" href="{}/css/Conf_Basic.css?{}">'.format(path, timestamp)
+
+        return '\n'.join([
+            css,
+            MathjaxMixin._getHeadContent(self),
+            WPEventBase._getHeadContent(self)
+        ])
+
+    def _applyDecoration(self, body):
+        self.logo_url = self.event.logo_url if self.event.has_logo else None
+        body = self._apply_conference_layout(body)
+        css_override_form = self._kwargs.get('css_override_form')
+        if css_override_form:
+            override_html = render_template('events/layout/css_preview_header.html',
+                                            event=self.event, form=css_override_form,
+                                            download_url=self._kwargs['css_url_override'])
+            body = override_html + body
+        return WPEventBase._applyDecoration(self, to_unicode(body))
 
 
-class WPConferenceDisplay(WPConferenceDefaultDisplayBase):
+class WPConferenceDisplay(WPConferenceDisplayLegacyBase):
     menu_entry_name = 'overview'
 
     def _getBody(self, params):
@@ -135,3 +292,10 @@ class WPConferenceDisplay(WPConferenceDefaultDisplayBase):
 
     def _getFooter(self):
         return render_event_footer(self.event).encode('utf-8')
+
+
+class WPAccessKey(WPJinjaMixin, WPDecorated):
+    template_prefix = 'events/'
+
+    def _getBody(self, params):
+        return self._getPageContent(params)
