+++ indico/modules/rb/views/user/event.py
@@ -14,8 +14,8 @@
 # You should have received a copy of the GNU General Public License
 # along with Indico; if not, see <http://www.gnu.org/licenses/>.
 
-from indico.legacy.webinterface.pages.conferences import WPConferenceModifBase
 from indico.legacy.webinterface.wcomponents import TabControl, WTabControl, WTemplated
+from indico.modules.events.management.views import WPEventManagementLegacy
 from indico.modules.rb.models.reservations import Reservation
 from indico.modules.rb.views.user.reservations import (WPRoomBookingBookingDetails, WPRoomBookingModifyBooking,
                                                        WPRoomBookingNewBookingConfirm,
@@ -25,11 +25,11 @@ from indico.modules.rb.views.user.rooms import WPRoomBookingRoomDetails
 from indico.web.flask.util import url_for
 
 
-class WPRoomBookingEventBase(WPConferenceModifBase):
+class WPRoomBookingEventBase(WPEventManagementLegacy):
     sidemenu_option = 'room_booking'
 
     def getJSFiles(self):
-        return WPConferenceModifBase.getJSFiles(self) + self._includeJSPackage('RoomBooking')
+        return WPEventManagementLegacy.getJSFiles(self) + self._includeJSPackage('RoomBooking')
 
     def _createTabCtrl(self):
         self._tabCtrl = TabControl()
